Moving clustering to a separate namespace
