+++ core/src/main/java/io/paradoxical/cassieq/clustering/election/LeadershipProvider.java
@@ -1,4 +1,4 @@
-package io.paradoxical.cassieq.election;
+package io.paradoxical.cassieq.clustering.election;
 
 import io.paradoxical.cassieq.model.LeadershipRole;
