+++ core/src/test/java/io/paradoxical/cassieq/unittests/ClusterTests.java
@@ -2,6 +2,7 @@ package io.paradoxical.cassieq.unittests;
 
 import com.hazelcast.core.HazelcastInstance;
 import io.paradoxical.cassieq.unittests.modules.HazelcastTestModule;
+import lombok.Cleanup;
 import org.junit.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -9,8 +10,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 public class ClusterTests {
     @Test
     public void test_hazelcasts(){
-        final HazelcastInstance instance1 = new HazelcastTestModule().getInstance();
-        final HazelcastInstance instance2 = new HazelcastTestModule().getInstance();
+        @Cleanup("shutdown") final HazelcastInstance instance1 = new HazelcastTestModule().getInstance();
+        @Cleanup("shutdown") final HazelcastInstance instance2 = new HazelcastTestModule().getInstance();
 
         assertThat(instance2.getCluster().getMembers().size()).isEqualTo(2);
